Move controllers module under internal
...to conform to the standard Go project layout:
https://github.com/golang-standards/project-layout?tab=readme-ov-file#internal
