+++ cmd/main.go
@@ -30,9 +30,9 @@ import (
 	"github.com/submariner-io/admiral/pkg/names"
 	admversion "github.com/submariner-io/admiral/pkg/version"
 	"github.com/submariner-io/submariner-operator/api/v1alpha1"
-	"github.com/submariner-io/submariner-operator/controllers/metrics"
-	"github.com/submariner-io/submariner-operator/controllers/servicediscovery"
-	"github.com/submariner-io/submariner-operator/controllers/submariner"
+	"github.com/submariner-io/submariner-operator/internal/controllers/metrics"
+	"github.com/submariner-io/submariner-operator/internal/controllers/servicediscovery"
+	"github.com/submariner-io/submariner-operator/internal/controllers/submariner"
 	"github.com/submariner-io/submariner-operator/pkg/crd"
 	"github.com/submariner-io/submariner-operator/pkg/gateway"
 	"github.com/submariner-io/submariner-operator/pkg/lighthouse"
