+++ internal/controllers/submariner/broker_controller_test.go
@@ -22,8 +22,8 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/submariner-io/submariner-operator/api/v1alpha1"
-	submarinerController "github.com/submariner-io/submariner-operator/controllers/submariner"
-	"github.com/submariner-io/submariner-operator/controllers/test"
+	submarinerController "github.com/submariner-io/submariner-operator/internal/controllers/submariner"
+	"github.com/submariner-io/submariner-operator/internal/controllers/test"
 	"github.com/submariner-io/submariner-operator/pkg/discovery/globalnet"
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
