+++ internal/controllers/submariner/globalnet_resources.go
@@ -24,8 +24,8 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/submariner-io/admiral/pkg/names"
 	"github.com/submariner-io/submariner-operator/api/v1alpha1"
-	"github.com/submariner-io/submariner-operator/controllers/apply"
-	"github.com/submariner-io/submariner-operator/controllers/metrics"
+	"github.com/submariner-io/submariner-operator/internal/controllers/apply"
+	"github.com/submariner-io/submariner-operator/internal/controllers/metrics"
 	"github.com/submariner-io/submariner-operator/pkg/httpproxy"
 	"github.com/submariner-io/submariner-operator/pkg/images"
 	opnames "github.com/submariner-io/submariner-operator/pkg/names"
