+++ internal/controllers/submariner/np_syncer_resources.go
