+++ internal/controllers/submariner/submariner_networkdiscovery.go
