+++ internal/controllers/submariner/submariner_suite_test.go
@@ -31,8 +31,8 @@ import (
 	"github.com/submariner-io/admiral/pkg/names"
 	"github.com/submariner-io/admiral/pkg/syncer/broker"
 	"github.com/submariner-io/submariner-operator/api/v1alpha1"
-	submarinerController "github.com/submariner-io/submariner-operator/controllers/submariner"
-	"github.com/submariner-io/submariner-operator/controllers/test"
+	submarinerController "github.com/submariner-io/submariner-operator/internal/controllers/submariner"
+	"github.com/submariner-io/submariner-operator/internal/controllers/test"
 	"github.com/submariner-io/submariner-operator/pkg/discovery/network"
 	opnames "github.com/submariner-io/submariner-operator/pkg/names"
 	submarinerv1 "github.com/submariner-io/submariner/pkg/apis/submariner.io/v1"
